+++ NEWS.md
@@ -3,6 +3,7 @@ Red Panda C++ Version 2.22
   - fix: Crash at startup when current problem in the problem set is connected with source file.
   - fix: Double-clicking on touchpad can't select current word.
   - fix: foreach-loops are not correctly parsed.
+  - fix: '^' is not correctly handled as operator.
 
 Red Panda C++ Version 2.21